Move ComputeShader to pygfx.utils for now
